adding register and login

+++ lushlyrics-webapp-django-main/main/views.py
@@ -1,11 +1,11 @@
 from django.http.response import HttpResponse
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect,get_object_or_404
 from django.contrib.auth.models import User
-from .models import playlist_user
 from django.urls.base import reverse
 from django.contrib.auth import authenticate,login,logout
 from youtube_search import YoutubeSearch
 import json
+from django.contrib.auth.decorators import login_required
 # import cardupdate
 
 
@@ -13,6 +13,7 @@ import json
 f = open('card.json', 'r')
 CONTAINER = json.load(f)
 
+@login_required
 def default(request):
     global CONTAINER
 
@@ -26,24 +27,26 @@ def default(request):
     return render(request, 'player.html',{'CONTAINER':CONTAINER, 'song':song})
 
 
-
+@login_required
 def playlist(request):
-    cur_user = playlist_user.objects.get(username = request.user)
-    try:
-      song = request.GET.get('song')
-      song = cur_user.playlist_song_set.get(song_title=song)
-      song.delete()
-    except:
-      pass
-    if request.method == 'POST':
-        add_playlist(request)
-        return HttpResponse("")
-    song = 'kSFJGEHDCrQ'
-    user_playlist = cur_user.playlist_song_set.all()
-    # print(list(playlist_row)[0].song_title)
+    user_playlist = request.user.playlist_songs.all()
+    song = request.GET.get('song')
+
+    if song:
+      try:
+        song = request.user.playlist_songs.get(song_title=song)
+        song.delete()
+      except:
+        pass
+      if request.method == 'POST':
+          add_playlist(request)
+          return HttpResponse("")
+      song = 'kSFJGEHDCrQ'
+      
+      # print(list(playlist_row)[0].song_title)
     return render(request, 'playlist.html', {'song':song,'user_playlist':user_playlist})
 
-
+@login_required
 def search(request):
   if request.method == 'POST':
 
@@ -61,9 +64,8 @@ def search(request):
 
 
 
-
 def add_playlist(request):
-    cur_user = playlist_user.objects.get(username = request.user)
+    cur_user = request.user.objects.get(username = request.user.username)
 
     if (request.POST['title'],) not in cur_user.playlist_song_set.values_list('song_title', ):
